Added projects support
added the ability to create multiple projects, each having separate
items and results. only restriction is that all projects have to have
the same item type `T`. i'll probably have to change the default root
from ~/.config/ranker, since this is supposed to be only a library, not
a fully fledged application.
also updated the `results` vector to have the same bitmask size as
`bitmasks`, essentially they both now are essentially the same datatype.
the previous implementation had the `results` more than twice as large
as `bitmasks`. this is supposed to speed up the parsing and caching, but
speed improvements are basically imperceptible, damn. next i'll have to
separate the ranker to its own standalone library, and then finally
start working on the actual TUI ranker application.