Moved `ranker` to a standalone library.
simplified the usage of the library: simplified `main`, added `load_projects` and `load_projects_from` with the default path being the cwd.
added an `ProjectsError` enum which holds extra information on why a
function errored.
the library now uses `debug!` and `info!` correctly, and decreased the
number of log messages overall.
added the ability to delete projects.