initial TUI ground work
finally implemented the ground work for the TUI: screens, popups, and
the image rendering backend, though this was ~stolen~ copied from
`moviedb`. i tried to avoid making the same mistakes i made in
`moviedb`, which means everything is modular and easily expandable. i'm
having doubts that the borrow checker is going to like this much
"freedom", so i expect major refactoring when i *actually* implement the
TUI, but thats for future me to deal with.