Fixes and polishing
the image backend now resizes each image it decodes and caches it for
future use, this speeds things up dramatically, and also makes
preloading window images significantly cheaper than before.
added a new dynamic popup method that adapts the size of the popup to
window area.
fixed a couple of bugs with the ranker lib.